requirements.txt
numpy==1.23.5
pillow==9.4.0
scikit-learn==1.2.1
tensorflow==2.12.0rc1
keras==2.12.0rc1
segmentation-models==1.0.1